rename package to match name in .toml
