+++ onboarding-enabler-python/requirements.txt
@@ -1,6 +1,23 @@
+aiohappyeyeballs==2.6.1
 aiohttp==3.10.10
+aiosignal==1.3.2
+attrs==25.3.0
+build==1.2.2.post1
+dnspython==2.7.0
+frozenlist==1.5.0
+idna==3.10
+ifaddr==0.2.0
+iniconfig==2.1.0
+multidict==6.4.3
+packaging==24.2
+pluggy==1.5.0
+propcache==0.3.1
 py_eureka_client==0.11.10
+pyproject_hooks==1.2.0
 pytest==8.3.5
-PyYAML==6.0.2
+pytest-asyncio==0.26.0
+pytest-mock==3.14.0
 PyYAML==6.0.2
 setuptools==75.2.0
+wheel==0.45.1
+yarl==1.19.0
